small export cleanup
