+++ fast64_internal/oot/cutscene/motion/exporter/__init__.py
@@ -1,1 +1,1 @@
-from .functions import exportCutsceneMotion
+from .functions import getCutsceneMotionData
