+++ fast64_internal/oot/cutscene/motion/exporter/functions.py
@@ -1,8 +1,32 @@
-from bpy.types import Context
-from .classes import OOTCutsceneMotionExport
-from .to_c import getCutsceneMotionData
+import bpy
+
+from bpy.types import Object
+from .classes import OOTCSMotionExport
 
 
-def exportCutsceneMotion(context: Context, filename: str, use_floats: bool, use_cscmd: bool):
-    exportData = OOTCutsceneMotionExport(context, use_floats, use_cscmd)
-    return exportData.exportToC(filename)
+def getCSMotionObjects():
+    csMotionObjects: dict[str, list[Object]] = {
+        "Cutscene": [],
+        "CS Actor Cue List": [],
+        "CS Player Cue List": [],
+        "CS Actor Cue": [],
+        "CS Player Cue": [],
+        "camShot": [],
+    }
+
+    for obj in bpy.data.objects:
+        if obj.type == "EMPTY" and obj.ootEmptyType in csMotionObjects.keys():
+            csMotionObjects[obj.ootEmptyType].append(obj)
+
+        if obj.type == "ARMATURE" and obj.parent.ootEmptyType == "Cutscene":
+            csMotionObjects["camShot"].append(obj)
+
+    return csMotionObjects
+
+
+def getCutsceneMotionData():
+    csMotionExport = OOTCSMotionExport(
+        getCSMotionObjects(), bpy.context.scene.fast64.oot.hackerFeaturesEnabled or bpy.context.scene.useDecompFeatures
+    )
+
+    return csMotionExport.getExportData()
