--- fast64_internal/oot/cutscene/motion/exporter/to_c/__init__.py
@@ -1,3 +1,0 @@
-from .motion import (
-    getCutsceneMotionData,
-)
